+++ lib/compendium-api.rb
@@ -1,5 +1,6 @@
 # Base Class
 require 'nretnil/compendium-api/compendium'
+require 'nretnil/compendium-public-api/compendium-public'
 # Endpoint Classes
 require 'nretnil/compendium-api/calendar'
 require 'nretnil/compendium-api/callback'
@@ -15,6 +16,8 @@ require 'nretnil/compendium-api/role'
 require 'nretnil/compendium-api/task'
 require 'nretnil/compendium-api/user'
 require 'nretnil/compendium-api/version'
+# Public Endpoint Classes
+require 'nretnil/compendium-public-api/content'
 # Helper Class
 require 'nretnil/compendium-helpers/helpers'
 require 'nretnil/compendium-helpers/json'