+++ lib/nretnil/compendium-public-api/content.rb
@@ -1,0 +1,35 @@
+module Nretnil
+  module CompendiumAPI
+
+    class CompendiumPublic
+
+      def content
+        ContentPublicAPI.new(self)
+      end
+
+    end
+
+    class ContentPublicAPI
+
+      def initialize(s)
+        @session = s
+      end
+
+      def list(publisher_id, options = {})
+        query = options
+        response = @session.get( '/api/publishers/' + publisher_id + '/feed', query )
+      end
+
+      def get(publisher_id, content_id, options = {})
+        query = options
+        response = @session.get( '/api/publishers/' + publisher_id + '/feed/' + content_id, query )
+      end
+
+      def required_params
+        @auth
+      end
+
+    end
+
+  end
+end